Choose display
I added the option to decide which fields to display.  The frame for presets is now functional so that there is only one window required.  Next I would like to build a class for races because I want to have the option for a user to add their own race.